♻️ Refactor : product.js 코드 리팩토링 - product.js 함수 분리 - createProductCard.js: 카드 생성부 - filter.js: 아코디언 필터 받아오는 부분 - sort.js: 정렬 버튼
